[COLAB-961] move ContestMassAction to entities
