+++ view/src/main/java/org/xcolab/view/pages/contestmanagement/entities/ContestMassAction.java
@@ -1,7 +1,7 @@
-package org.xcolab.view.pages.contestmanagement.entities.massactions;
+package org.xcolab.view.pages.contestmanagement.entities;
 
 import org.xcolab.client.contest.pojo.Contest;
-import org.xcolab.view.pages.contestmanagement.entities.MassActionRequiresConfirmationException;
+import org.xcolab.view.pages.contestmanagement.entities.massactions.MassActionDataWrapper;
 
 import java.io.IOException;
 import java.util.List;
