+++ view/src/main/java/org/xcolab/view/pages/contestmanagement/entities/ContestMassActions.java
@@ -1,7 +1,6 @@
 package org.xcolab.view.pages.contestmanagement.entities;
 
 import org.xcolab.view.pages.contestmanagement.entities.massactions.ActiveMassAction;
-import org.xcolab.view.pages.contestmanagement.entities.massactions.ContestMassAction;
 import org.xcolab.view.pages.contestmanagement.entities.massactions.DeleteMassAction;
 import org.xcolab.view.pages.contestmanagement.entities.massactions.DeleteWithPhasesMassAction;
 import org.xcolab.view.pages.contestmanagement.entities.massactions.FeaturedMassAction;
